a) New basic processing functions b) New graphs of performance c) .gitignore changes
Some functions to get specific data from issues.

Added two plots to analyze performance in a sprint.

Lots of white space indenting. For not showing those unimportant
changes in git log, use the -w or --ignore-all-space option

Change in .gitignore to include the *.tgz with the sprint data.
Was needed to clean the stage area at the root of the repo:
git rm -r --cached .
git add -A